Rewriting Gallery Viewer in Cornerstone
