+++ src/components/detailedView/displays/XtkViewer/XtkViewer.tsx
@@ -1,193 +1,202 @@
-import React, { useEffect, useRef, useState } from 'react'
-import { Button } from '@patternfly/react-core'
-import { AiOutlineExpand } from 'react-icons/ai'
-import { FeedFile } from '@fnndsc/chrisapi'
-import { useTypedSelector } from '../../../../store/hooks'
-import FsmFileSelect from './PrimaryFileSelect'
-import CrvFileSelect from './CrvFileSelect'
-import './xtk-viewer.scss'
+import React, { useEffect, useRef, useState } from "react";
+import { Button } from "@patternfly/react-core";
+import { AiOutlineExpand } from "react-icons/ai";
+import { FeedFile } from "@fnndsc/chrisapi";
+import { useTypedSelector } from "../../../../store/hooks";
+import FsmFileSelect from "./PrimaryFileSelect";
+import CrvFileSelect from "./CrvFileSelect";
+import "./xtk-viewer.scss";
 
 // X and dat are loaded in from script file
-declare const X: any
-declare const dat: any
+declare const X: any;
+declare const dat: any;
 
-export type ViewerMode = 'volume' | 'mesh' | 'other'
-type VolumeMode = '3D' | '2D'
+export type ViewerMode = "volume" | "mesh" | "other";
+type VolumeMode = "3D" | "2D";
 
 const getFileType = (file?: FeedFile) =>
-  file?.data.fname.split('.').slice(-1)[0]
+  file?.data.fname.split(".").slice(-1)[0];
 const getFileData = async (file: FeedFile) =>
-  (await file.getFileBlob())?.arrayBuffer()
+  (await file.getFileBlob())?.arrayBuffer();
 
 export function getXtkFileMode(fileType?: string): ViewerMode | undefined {
-  const volumeExtensions = ['mgz', 'dcm']
-  const meshExtensions = ['fsm', 'smoothwm', 'pial']
-  const otherExtensions = ['crv']
+  const volumeExtensions = ["mgz", "dcm"];
+  const meshExtensions = ["fsm", "smoothwm", "pial"];
+  const otherExtensions = ["crv"];
   if (!fileType) {
-    return
+    return;
   } else if (volumeExtensions.includes(fileType)) {
-    return 'volume'
+    return "volume";
   } else if (meshExtensions.includes(fileType)) {
-    return 'mesh'
+    return "mesh";
   } else if (otherExtensions.includes(fileType)) {
-    return 'other'
+    return "other";
   }
 }
 
 function getPrimaryFileMode(file: FeedFile): ViewerMode | undefined {
-  const fileType = getFileType(file)
-  return getXtkFileMode(fileType)
+  const fileType = getFileType(file);
+  return getXtkFileMode(fileType);
 }
 
 const XtkViewer = () => {
   const directoryFiles =
-    useTypedSelector((state) => state.explorer.selectedFolder) || []
-  const crvFiles = directoryFiles.filter((file) => {
-    const fileName = file.data.fname
-    return fileName?.endsWith('.crv')
-  })
+    useTypedSelector((state) => state.explorer.selectedFolder) || [];
 
-  const selectedFile = useTypedSelector((state) => state.explorer.selectedFile)
-  const selectedFileType = getFileType(selectedFile)
+  const crvFiles = directoryFiles.filter((file) => {
+    const fileName = file.data.fname;
+    return fileName?.endsWith(".crv");
+  });
+
+  const selectedFile = useTypedSelector((state) => state.explorer.selectedFile);
+  const selectedFileType = getFileType(selectedFile);
 
   // data based on first load, before further interactions
   const defaultPrimaryFile =
-    selectedFileType === 'crv' ? undefined : selectedFile
-  const defaultCrvFile = selectedFileType === 'crv' ? selectedFile : undefined
+    selectedFileType === "crv" ? undefined : selectedFile;
+  const defaultCrvFile = selectedFileType === "crv" ? selectedFile : undefined;
   const defaultViewerMode = defaultPrimaryFile
     ? getPrimaryFileMode(defaultPrimaryFile)
-    : undefined
+    : undefined;
 
   const [primaryFile, setPrimaryFile] = useState<FeedFile | undefined>(
-    defaultPrimaryFile,
-  )
+    defaultPrimaryFile
+  );
   const [viewerMode, setViewerMode] = useState<ViewerMode | undefined>(
-    defaultViewerMode,
-  )
-  const [volumeMode, setVolumeMode] = useState<VolumeMode>('3D')
-  const [crvFile, setCrvFile] = useState<FeedFile | undefined>(defaultCrvFile)
-  const [secondaryFile, setSecondaryFile] = useState<FeedFile | undefined>()
-  const [orientation, setOrientation] = useState('x')
-  const [fullscreen, setFullscreen] = useState(false)
+    defaultViewerMode
+  );
+  const [volumeMode, setVolumeMode] = useState<VolumeMode>("3D");
+  const [crvFile, setCrvFile] = useState<FeedFile | undefined>(defaultCrvFile);
+  const [secondaryFile, setSecondaryFile] = useState<FeedFile | undefined>();
+  const [orientation, setOrientation] = useState("x");
+  const [fullscreen, setFullscreen] = useState(false);
 
   useEffect(() => {
-    let r: any
-    let gui: any
-    let object: any
-    let secondaryObject: any
+    let r: any;
+    let gui: any;
+    let object: any;
+    let secondaryObject: any;
 
     async function renderFileData() {
       if (!primaryFile || !viewerMode) {
-        return
+        return;
       }
 
-      const fileData = await getFileData(primaryFile)
+      const fileData = await getFileData(primaryFile);
 
       if (!fileData) {
-        return
+        return;
       }
 
-      if (viewerMode === 'volume') {
-        if (volumeMode === '3D') {
-          r = new X.renderer3D()
+      if (viewerMode === "volume") {
+        if (volumeMode === "3D") {
+          r = new X.renderer3D();
         } else {
-          r = new X.renderer2D()
-          r.orientation = orientation
+          r = new X.renderer2D();
+          r.orientation = orientation;
         }
 
-        object = new X.volume()
-        ;(object.file = primaryFile.data.fname), (object.filedata = fileData)
+        object = new X.volume();
+        object.file = primaryFile.data.fname;
+        object.filedata = fileData;
       }
 
-      if (viewerMode === 'mesh') {
-        r = new X.renderer3D()
-        object = new X.mesh()
-        object.file = primaryFile.data.fname
-        object.filedata = fileData
+      if (viewerMode === "mesh") {
+        r = new X.renderer3D();
+        object = new X.mesh();
+        object.file = primaryFile.data.fname;
+        object.filedata = fileData;
         if (crvFile) {
-          object.scalars.file = 'crv_file.crv'
-          object.scalars.filedata = await getFileData(crvFile)
+          object.scalars.file = "crv_file.crv";
+          object.scalars.filedata = await getFileData(crvFile);
         }
 
         if (secondaryFile) {
-          secondaryObject = new X.mesh()
-          secondaryObject.file = secondaryFile.data.fname
-          const secondaryFileData = await getFileData(secondaryFile)
-          secondaryObject.filedata = secondaryFileData
+          secondaryObject = new X.mesh();
+          secondaryObject.file = secondaryFile.data.fname;
+          const secondaryFileData = await getFileData(secondaryFile);
+          secondaryObject.filedata = secondaryFileData;
         }
       }
 
-      r.container = renderContainerRef.current
-      r.init()
-      r.camera.position = [0, 400, 0]
+      r.container = renderContainerRef.current;
+      r.init();
+      r.camera.position = [0, 400, 0];
 
-      r.add(object)
+      r.add(object);
       if (secondaryObject) {
-        r.add(secondaryObject)
+        r.add(secondaryObject);
       }
 
       r.onShowtime = function () {
-        gui = new dat.GUI()
+        gui = new dat.GUI();
 
-        if (viewerMode === 'mesh') {
-          const meshgui = gui.addFolder('Mesh')
-          meshgui.addColor(object, 'color')
-          meshgui.open()
+        if (viewerMode === "mesh") {
+          const meshgui = gui.addFolder("Mesh");
+          meshgui.addColor(object, "color");
+          meshgui.open();
         }
 
         if (crvFile) {
-          const curvgui = gui.addFolder('Curvature')
-          const { scalars } = object
-          curvgui.addColor(scalars, 'minColor')
-          curvgui.addColor(scalars, 'maxColor')
-          curvgui.add(scalars, 'lowerThreshold', scalars.min, scalars.max)
-          curvgui.add(scalars, 'upperThreshold', scalars.min, scalars.max)
-          curvgui.open()
+          const curvgui = gui.addFolder("Curvature");
+          const { scalars } = object;
+          curvgui.addColor(scalars, "minColor");
+          curvgui.addColor(scalars, "maxColor");
+          curvgui.add(scalars, "lowerThreshold", scalars.min, scalars.max);
+          curvgui.add(scalars, "upperThreshold", scalars.min, scalars.max);
+          curvgui.open();
         }
 
-        if (viewerMode === 'volume') {
-          const volumegui = gui.addFolder('Volume')
-          volumegui.add(object, 'indexX', 0, 248)
-          volumegui.add(object, 'indexY', 0, 248)
-          volumegui.add(object, 'indexZ', 0, 248)
-          volumegui.open()
+        if (viewerMode === "volume") {
+          const volumegui = gui.addFolder("Volume");
+          volumegui.add(object, "indexX", 0, 248);
+          volumegui.add(object, "indexY", 0, 248);
+          volumegui.add(object, "indexZ", 0, 248);
+          volumegui.open();
         }
 
         // manually move controls to inside panel
-        datGuiContainerRef.current?.appendChild(gui.domElement.parentElement)
-      }
+        datGuiContainerRef.current?.appendChild(gui.domElement.parentElement);
+      };
 
-      r.render()
+      r.render();
     }
 
-    renderFileData()
+    renderFileData();
 
     return () => {
       try {
-        r.destroy()
-        gui.destroy()
+        r.destroy();
+        gui.destroy();
       } catch (e) {
-        console.log(e)
+        console.log(e);
       }
-    }
-  }, [viewerMode, volumeMode, primaryFile, crvFile, orientation, secondaryFile])
+    };
+  }, [
+    viewerMode,
+    volumeMode,
+    primaryFile,
+    crvFile,
+    orientation,
+    secondaryFile,
+  ]);
 
   const handleFullscreenToggle = () => {
     if (!fullscreen) {
-      fullscreenRef?.current?.requestFullscreen()
+      fullscreenRef?.current?.requestFullscreen();
     } else if (document.fullscreenElement) {
-      document.exitFullscreen()
+      document.exitFullscreen();
     }
-    setFullscreen(!fullscreen)
-  }
+    setFullscreen(!fullscreen);
+  };
 
-  const renderContainerRef = useRef(null)
-  const datGuiContainerRef = useRef<HTMLDivElement>(null)
-  const fullscreenRef = useRef<HTMLDivElement>(null)
+  const renderContainerRef = useRef(null);
+  const datGuiContainerRef = useRef<HTMLDivElement>(null);
+  const fullscreenRef = useRef<HTMLDivElement>(null);
 
   return (
     <div
-      className={`xtk-viewer-wrap ${fullscreen ? 'fullscreen' : ''}`}
+      className={`xtk-viewer-wrap ${fullscreen ? "fullscreen" : ""}`}
       ref={fullscreenRef}
     >
       <div className="xtk-header">
@@ -210,8 +219,8 @@ const XtkViewer = () => {
             <div className="instructions">
               Please select a viewer mode for <u>{primaryFile?.data.fname}</u>
             </div>
-            <Button onClick={() => setViewerMode('volume')}>Volume</Button>
-            <Button onClick={() => setViewerMode('mesh')}>Mesh</Button>
+            <Button onClick={() => setViewerMode("volume")}>Volume</Button>
+            <Button onClick={() => setViewerMode("mesh")}>Mesh</Button>
           </div>
         )
       }
@@ -220,15 +229,15 @@ const XtkViewer = () => {
         <FsmFileSelect
           files={directoryFiles}
           handleSelect={(file) => {
-            setPrimaryFile(file)
-            setViewerMode(getPrimaryFileMode(file))
+            setPrimaryFile(file);
+            setViewerMode(getPrimaryFileMode(file));
           }}
         />
       ) : (
         <>
           <div ref={renderContainerRef} className="xtk-render"></div>
           <div className="xtk-controls">
-            {viewerMode === 'mesh' ? (
+            {viewerMode === "mesh" ? (
               <div className="additional-files">
                 <CrvFileSelect
                   files={crvFiles}
@@ -247,37 +256,37 @@ const XtkViewer = () => {
                 <div>
                   <div className="caption">&nbsp;</div>
                   <Button
-                    className={volumeMode === '3D' ? 'active' : ''}
-                    onClick={() => setVolumeMode('3D')}
+                    className={volumeMode === "3D" ? "active" : ""}
+                    onClick={() => setVolumeMode("3D")}
                   >
                     3D
                   </Button>
                   <Button
-                    className={volumeMode === '2D' ? 'active' : ''}
-                    onClick={() => setVolumeMode('2D')}
+                    className={volumeMode === "2D" ? "active" : ""}
+                    onClick={() => setVolumeMode("2D")}
                   >
                     2D
                   </Button>
                 </div>
-                {volumeMode === '2D' && (
+                {volumeMode === "2D" && (
                   <div className="orientation">
                     <div className="caption">Orientation</div>
                     <div className="orientation-button-wrap">
                       <Button
-                        className={orientation === 'x' ? 'active' : ''}
-                        onClick={() => setOrientation('x')}
+                        className={orientation === "x" ? "active" : ""}
+                        onClick={() => setOrientation("x")}
                       >
                         X
                       </Button>
                       <Button
-                        className={orientation === 'y' ? 'active' : ''}
-                        onClick={() => setOrientation('y')}
+                        className={orientation === "y" ? "active" : ""}
+                        onClick={() => setOrientation("y")}
                       >
                         Y
                       </Button>
                       <Button
-                        className={orientation === 'z' ? 'active' : ''}
-                        onClick={() => setOrientation('z')}
+                        className={orientation === "z" ? "active" : ""}
+                        onClick={() => setOrientation("z")}
                       >
                         Z
                       </Button>
@@ -295,7 +304,7 @@ const XtkViewer = () => {
         </>
       )}
     </div>
-  )
-}
+  );
+};
 
-export default XtkViewer
+export default XtkViewer;
