--- src/components/dicomViewer/DcmHeader/DcmHeader.scss
@@ -1,106 +1,0 @@
-@import "../../../assets/scss/helpers/mixins.scss";
-
-$imgViewerBackgroundColour: black;
-$textColor: white;
-$GreyTextColor: #868686;
-$opacityLevel: 0.75;
-$headerBlue: #73bcf7;
-$iconfont-size-md: 1rem;
-$iconfont-size-lg: 1.2rem;
-
-.dicom-header {
-  z-index: 10;
-  padding-bottom:0;
-  background: rgba(0, 0, 0, $opacityLevel); //$imgViewerBackgroundColour;
-  color: $textColor;
-  display: flex;
-  flex-wrap: nowrap;
-  flex-direction: row;
-  justify-content: center;
-
-
-
-  .dicom-logo {
-    position: absolute;
-    left: 0;
-    top: 0;
-    margin-right: 4rem;
-    line-height: 1.1;
-    margin-left: 1rem;
-    a {
-      color: #fff;
-    }
-
-    &__text {
-      @include media-query(1200px) {
-        font-size: 0.8rem;
-      }
-
-      @include media-query(1920px) {
-        font-size: 1.4rem;
-      }
-
-      color: $headerBlue !important;
-      font-weight: 900;
-      display: block;
-    }
-  }
-
-  button {
-    margin: 0.4rem -4px;
-    svg {
-      color: #fff;
-      font-size: 1.5rem !important;
-
-      @include media-query(1200px) {
-        font-size: $iconfont-size-md;
-      }
-
-      @include media-query(1920px) {
-        font-size: $iconfont-size-lg;
-      }
-    }
-  }
-
-  .dicom-header__center-tools {
-    display: flex;
-    flex-flow: row nowrap;
-
-    .dicom-header__tools-text {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      @include media-query(1200px) {
-        font-size: $iconfont-size-lg;
-      }
-
-      @include media-query(1920px) {
-        font-size: 1rem;
-      }
-    }
-  }
-
-  .pf-m-active {
-    color: red !important;
-  }
-
-  button:hover,
-  button:focus {
-    svg {
-      color: $headerBlue;
-    }
-  }
-}
-
-.dicomTag {
-  min-height: calc(100vh - 400px);
-  &__list {
-    &__item {
-      font-family: monospace;
-      &--name {
-        font-weight: bold;
-      }
-    }
-  }
-}
