--- src/components/dicomViewer/amiViewer.scss
@@ -1,53 +1,0 @@
-/*
-*  File:            amiViewer.scss
-*  Description:     This file holds styles specific for the dcm viewer
-*  Author:          ChRIS UI
-*/
-
-.ami-viewer {
-  position: relative;
-  min-height: calc(100vh - 300px);
-  #container {
-    bottom: 0;
-    height: 100%;
-    left: 0;
-    overflow: hidden;
-    position: absolute;
-    right: 0;
-    top: 0;
-    width: 100%;
-    font-size: 0.7rem;
-  }
-  .iframe-container {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-
-    iframe {
-      height: 100% !important;
-    }
-  }
-}
-
-@media (min-width: 1020px) {
-  .ami-viewer {
-    #container {
-      font-size: 1rem;
-    }
-  }
-}
-
-.loader {
-  margin-top: -5rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  flex-direction: column;
-  gap: 1.5rem;
-  background: black;
-
-  span {
-    color: white;
-  }
-}
