+++ src/components/feed/FeedOutputBrowser/FeedOutputBrowser.scss
@@ -162,4 +162,15 @@
   }
 }
 
+.overlay {
+  /* prevent text selection on overlay */
+  -webkit-touch-callout: none;
+  -webkit-user-select: none;
+  -khtml-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
 
+  /* ignore pointer event on overlay */
+  pointer-events: none;
+}
