--- src/components/gallery/GalleryArrows/GalleryArrows.scss
@@ -1,27 +1,0 @@
-/*
-*  File:            GalleryArrows.scss
-*  Description:     This files holds styles specific for the file explorer Gallery Arrows
-*  Author:          ChRIS UI
-*/
-
-@import "../../../assets/scss/helpers/variables";
-
-.arrows> a {
-    color: #fff;
-    position: absolute;
-    z-index: 9999;
-  }
-  // Naviation Arrows
-  .arrows{
-    a{
-      font-size: $gallery-arrows-height;
-      margin-top: -($gallery-arrows-height*0.55);
-      top: 50%;
-      &.prev{
-        left: 0;
-      }
-      &.next{
-        right: 0;
-      }
-    }
-  }
