--- src/components/gallery/GalleryInfoPanel/GalleryInfoPanel.scss
@@ -1,17 +1,0 @@
-/*
-*  File:            GalleryInfoPanel.scss
-*  Description:     This files holds styles specific for the file explorer Gallery Toolbar
-*  Author:          ChRIS UI
-*/
-@import "../../../assets/scss/helpers/variables";
-.gallery-info{
-  background: rgba(0, 0, 0, 0.5);
-  color: #fff;
-  font-size: 0.875rem;
-  left: 0;
-  padding: 1rem;
-  position: absolute;
-  top: 0;
-  width: 100%;
-  z-index: 9999;
-}
