--- src/components/gallery/GalleryWrapper.scss
@@ -1,26 +1,0 @@
-/*
-*  File:            GalleryWrapper.scss
-*  Description:     This files holds styles specific for the file explorer Gallery 
-*  Author:          ChRIS UI
-*/
-
-@import "../../assets/scss/helpers/variables";
-
-.gallery-wrapper {
-  background: black;
-  border: solid 1px $fe-viewer-border;
-  min-height: calc(100vh - 250px);
-  // padding: 1rem 3rem 3rem;
-  overflow: hidden;
-  position: relative;
-  max-width: 95vw;
-  button.close-btn {
-    position: absolute;
-    right: 0;
-    top: 10px;
-    //z-index: 10;
-    svg {
-      color: #fff;
-    }
-  }
-}
