--- src/components/gallery/index.ts
@@ -1,4 +1,0 @@
-export  { default as GalleryWrapper} from "./GalleryWrapper";
-export  { default as GalleryArrows} from "./GalleryArrows/GalleryArrows";
-export  { default as GalleryToolbar} from "./GalleryToolbar/GalleryToolbar";
-export  { default as GalleryFullScreen} from "./GalleryFullScreen/GalleryFullScreen";
